fix: dm category mapping
